requirements.txt
streamlit>=1.32.0
langchain-community>=0.1.0
langchain-core>=0.1.0
langchain-experimental>=0.0.55
pydantic>=2.0.0,<3.0.0
nomic>=3.0.0
neo4j>=5.0.0
pyvis>=0.3.0
networkx>=3.0
pillow>=10.0.0